requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
flask-cors==3.0.10
cachetools==5.5.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
gunicorn==21.2.0  # Required for deployment on Render
